make a duplicate of forms when duplicating proposal templates
