+++ lib/proposals/__tests__/createProposal.spec.ts
@@ -78,107 +78,6 @@ describe('Creates a page and proposal with relevant configuration', () => {
     );
   });
 
-  it('Create proposal template with existing form', async () => {
-    const reviewerUser = await generateSpaceUser({
-      isAdmin: false,
-      spaceId: space.id
-    });
-    const extraUser = await generateSpaceUser({
-      isAdmin: false,
-      spaceId: space.id
-    });
-
-    const pageTitle = 'page title 124';
-
-    const formFields: FormFieldInput[] = [
-      {
-        id: uuid(),
-        type: 'short_text',
-        name: 'name',
-        description: 'description',
-        index: 0,
-        options: [],
-        private: false,
-        required: true,
-        fieldConfig: {}
-      },
-      {
-        id: uuid(),
-        type: 'long_text',
-        name: 'long name',
-        description: 'another description',
-        index: 1,
-        options: [],
-        private: true,
-        required: true,
-        fieldConfig: {}
-      }
-    ];
-
-    const proposalTemplate = await testUtilsProposals.generateProposal({
-      spaceId: space.id,
-      userId: user.id
-    });
-
-    const proposalTemplateForm = await prisma.form.create({
-      data: {
-        proposal: {
-          connect: {
-            id: proposalTemplate.id
-          }
-        }
-      }
-    });
-
-    await prisma.formField.createMany({
-      data: formFields.map((item) => ({
-        ...item,
-        description: item.description ?? '',
-        formId: proposalTemplateForm.id,
-        fieldConfig: {}
-      }))
-    });
-
-    const { page, proposal } = await createProposal({
-      pageProps: {
-        contentText: '',
-        title: pageTitle,
-        type: 'proposal_template'
-      },
-      userId: user.id,
-      spaceId: space.id,
-      authors: [user.id, extraUser.id],
-      evaluations: [],
-      formFields: formFields.map((item) => ({
-        ...item,
-        // creating new IDs for the form fields
-        id: v4()
-      })),
-      workflowId: workflow.id
-    });
-
-    const newProposalTemplateForm = await prisma.form.findUniqueOrThrow({
-      where: {
-        id: proposal.formId as string
-      },
-      select: {
-        id: true
-      }
-    });
-
-    // New form should be created since we are duplicating a proposal template
-    expect(newProposalTemplateForm.id).not.toEqual(proposalTemplateForm.id);
-
-    expect(page).toMatchObject(
-      expect.objectContaining({
-        title: pageTitle,
-        type: 'proposal_template'
-      })
-    );
-
-    expect(proposal.formId).toBeDefined();
-  });
-
   it('should create a proposal from a workflow and copy over its permissions', async () => {
     const evaluationTemplate: WorkflowEvaluationJson[] = [
       {
